+++ RELEASENOTES.md
@@ -1,5 +1,20 @@
 # Release notes of ion-autocomplete
 
+## Version 0.3.1
+
+* Tag: [0.3.1](https://github.com/guylabs/ion-autocomplete/tree/v0.3.1)
+* Release: [ion-autocomplete-0.3.1.zip](https://github.com/guylabs/ion-autocomplete/archive/v0.3.1.zip)
+
+### Changes
+
+* Added possibility to pre populate the search items - [#57](https://github.com/guylabs/ion-autocomplete/issues/57).
+* Added the `max-selected-items` attribute to restrict the selected search items - [#58](https://github.com/guylabs/ion-autocomplete/issues/58).
+
+### Migration notes
+
+* As of version `0.3.1` the `multiple-select` attribute has been dropped in favor of the `max-selected-items` attribute.
+Please have a look at the documentation here https://github.com/guylabs/ion-autocomplete#the-max-selected-items on how to migrate this.
+
 ## Version 0.3.0
 
 * Tag: [0.3.0](https://github.com/guylabs/ion-autocomplete/tree/v0.3.0)
@@ -14,7 +29,7 @@
 
 ### Migration notes
 
-* As of version 0.3.0 the component does not support the element restriction anymore, such that you are just able to 
+* As of version `0.3.0` the component does not support the element restriction anymore, such that you are just able to 
 use the attribute restriction on all your elements. This means that you need to convert all `<ion-autocomplete ... />` 
  tags to the following tag: `<input ion-autocomplete type="text" readonly="readonly" class="ion-autocomplete" autocomplete="off" ... />`